--- .github/workflows/main.yml
@@ -1,34 +1,0 @@
-name: Hourly News Bot
-on:
-  schedule:
-    - cron: '17 * * * *' # 每小時執行一次
-  workflow_dispatch:   # 允許手動點擊執行
-
-jobs:
-  run-bot:
-    runs-on: ubuntu-latest
-    permissions:
-      contents: write # 允許它把報表存回 GitHub
-    steps:
-      - uses: actions/checkout@v4
-      - name: Setup Node
-        uses: actions/setup-node@v4
-        with:
-          node-version: '20'
-      - name: Install
-        run: npm install
-      - name: Run
-        env:
-          GEMINI_API_KEY: ${{ secrets.GEMINI_API_KEY }}
-          DISCORD_WEBHOOK_URL: ${{ secrets.DISCORD_WEBHOOK_URL }}
-          NEWS_SOURCES: ${{ secrets.NEWS_SOURCES }}
-          KEYWORDS: ${{ secrets.KEYWORDS }}
-          EXCLUDE_KEYWORDS: ${{ secrets.EXCLUDE_KEYWORDS }}
-        run: node index.js
-      - name: Save Data
-        run: |
-          git config --local user.name "github-actions[bot]"
-          git config --local user.email "github-actions[bot]@users.noreply.github.com"
-          git add news_bot.db reports/
-          git commit -m "Auto update news data" || echo "No changes"
-          git push
