--- .github/workflows/weekly.yml
@@ -1,28 +1,0 @@
-name: Weekly Deep Dive
-
-on:
-  schedule:
-    - cron: '0 14 * * 0' # 每週日晚上 10:00 (台灣時間)
-  workflow_dispatch: # 保留手動觸發按鈕
-
-jobs:
-  weekly-report:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-      
-      - name: Setup Node.js
-        uses: actions/setup-node@v4
-        with:
-          node-version: '20'
-          # 這裡移除了 cache: 'npm'，避免報錯
-
-      - name: Install dependencies
-        run: npm install
-
-      - name: Run Weekly Analysis
-        env:
-          GEMINI_API_KEY: ${{ secrets.GEMINI_API_KEY }}
-          DISCORD_WEBHOOK_URL: ${{ secrets.DISCORD_WEBHOOK_URL }}
-          NEWS_SOURCES: ${{ secrets.NEWS_SOURCES }}
-        run: node weekly.js
