--- app.js
@@ -1,10 +1,0 @@
-require('dotenv').config();
-const { GoogleGenerativeAI } = require("@google/generative-ai");
-const Parser = require('rss-parser');
-const axios = require('axios');
-const { JSDOM } = require('jsdom');
-const { Readability } = require('@mozilla/readability');
-
-// 初始化 Gemini
-const genAI = new GoogleGenerativeAI(process.env.GEMINI_API_KEY);
-const parser = new Parser();
