+++ lib/ui.js
@@ -1,15 +1,14 @@
 const fs = require('fs');
 const path = require('path');
+const { publicDir } = require('./config');
+const { ensureDir } = require('./utils');
 
-/**
- * 專門處理 HTML 報表生成的模組 (含分類篩選功能)
- */
 function generateHTMLReport(summary, newsData) {
     const dateStr = new Date().toLocaleDateString('zh-TW', { 
         year: 'numeric', month: 'long', day: 'numeric', weekday: 'long' 
     });
     
-    // UI 格式化邏輯
+    // UI 格式化邏輯 (縮減版，詳細樣式同原版)
     const formattedSummary = summary
         .replace(/\n/g, '<br>')
         .replace(/🟢/g, '<span class="text-2xl animate-pulse">🟢</span> <b class="text-green-600">利多趨勢</b>')
@@ -18,7 +17,7 @@ function generateHTMLReport(summary, newsData) {
         .replace(/\*\*(.*?)\*\*/g, '<b class="text-indigo-600 font-bold">$1</b>')
         .replace(/### (.*?)(<br>|$)/g, '<h3 class="text-lg font-bold text-slate-800 mt-6 mb-3 border-l-4 border-indigo-500 pl-3 bg-indigo-50/50 py-1">$1</h3>');
 
-    // 生成分類篩選按鈕
+    // ... (Filter Buttons HTML 同原版，此處省略以節省空間) ...
     const categories = ["全部", "科技", "金融", "社會", "其他"];
     const filterButtonsHtml = categories.map(cat => `
         <button onclick="filterCategory('${cat}')" 
@@ -28,7 +27,7 @@ function generateHTMLReport(summary, newsData) {
         </button>
     `).join('');
 
-    // 生成新聞卡片
+    // ... (Cards HTML 同原版) ...
     const articlesHtml = newsData.map(n => `
         <div class="news-card group bg-white rounded-2xl p-6 shadow-sm hover:shadow-xl transition-all duration-300 border border-slate-100 flex flex-col justify-between" 
              data-category="${n.category || '其他'}">
@@ -52,6 +51,7 @@ function generateHTMLReport(summary, newsData) {
             </div>
         </div>`).join('');
 
+    // 完整的 HTML 結構
     const html = `<!DOCTYPE html>
     <html lang="zh-TW">
     <head>
@@ -69,7 +69,6 @@ function generateHTMLReport(summary, newsData) {
             function filterCategory(cat) {
                 const cards = document.querySelectorAll('.news-card');
                 const btns = document.querySelectorAll('.filter-btn');
-                
                 btns.forEach(btn => {
                     if (btn.innerText.trim() === cat) {
                         btn.className = 'filter-btn px-5 py-2 rounded-full text-sm font-bold border transition-all duration-300 transform active:scale-95 shadow-md bg-indigo-600 text-white border-indigo-600 shadow-indigo-300';
@@ -77,7 +76,6 @@ function generateHTMLReport(summary, newsData) {
                         btn.className = 'filter-btn px-5 py-2 rounded-full text-sm font-bold border transition-all duration-300 transform active:scale-95 shadow-sm bg-white text-slate-600 border-slate-200 hover:border-indigo-400 hover:text-indigo-600';
                     }
                 });
-
                 cards.forEach(card => {
                     card.classList.remove('animate-fade');
                     if (cat === '全部' || card.dataset.category === cat) {
@@ -123,12 +121,9 @@ function generateHTMLReport(summary, newsData) {
     </body>
     </html>`;
 
-    const reportsDir = './reports';
-    if (!fs.existsSync(reportsDir)) fs.mkdirSync(reportsDir);
-
-    // 🟢 關鍵修正：固定檔名為 index.html
+    ensureDir(publicDir);
     const fileName = 'index.html'; 
-    const filePath = path.join(reportsDir, fileName);
+    const filePath = path.join(publicDir, fileName);
     fs.writeFileSync(filePath, html);
 
     return { filePath, fileName };
