--- reports/report_2026-02-12.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html><html lang="zh-TW"><head><meta charset="UTF-8"><script src="https://cdn.tailwindcss.com"></script></head><body class="bg-slate-50 p-6"><div class="max-w-3xl mx-auto"><h1 class="text-3xl font-black mb-6 text-slate-800 text-center">AI 新聞去重日報</h1><div class="bg-white p-8 rounded-3xl shadow-xl mb-10 border-t-8 border-indigo-600"><h2>🤖 深度摘要</h2><div>身為專業分析師，我將對您提供的兩則新聞進行分類與深度摘要。<br><br>---<br><br>## 新聞分類與深度摘要<br><br>### 1. 新聞分類<br><br>兩則新聞皆屬於<strong>財經</strong>類別，具體聚焦於<strong>股市動態</strong>及<strong>科技產業影響</strong>。<br><br>### 2. 深度摘要<br><br>#### 總體市場表現與主要影響因素：<br><br>*   <strong>市場下挫：</strong> 兩則新聞都指出，週四（2026年2月13日）華爾街股市出現顯著下跌。道瓊工業指數下跌超過600點（具體為669.42點，跌幅1.34%），標準普爾500指數下跌1.57%，納斯達克綜合指數下跌2.03%，費城半導體指數下跌2.50%。<br>*   <strong>AI 顛覆性擔憂：</strong> 這是導致股市下跌的最主要原因。投資人開始擔憂人工智慧（AI）的快速發展可能顛覆現有產業的商業模式，引發失業率上升，從而對企業盈利和整體經濟前景產生負面影響。<br>*   <strong>科技股受創：</strong> 由於 AI 相關擔憂，科技股成為此次下跌的主要受災戶。<br>    *   <strong>台積電 ADR：</strong> 兩則新聞都提及，台積電的美國存託憑證（ADR）下跌了1.6%。<br>    *   <strong>其他科技巨頭：</strong> 第二則新聞更具體列出了多檔科技股的跌幅，包括輝達（-1.6%）、蘋果（-5%）、亞馬遜（-2.25%）、微軟（-0.63%）、Alphabet（-0.6%）、Meta（-2.82%）、特斯拉（-2.69%）。<br>    *   <strong>思科系統：</strong> 思科系統因發布平淡的財報，股價重挫逾12%，進一步加劇了科技股的低迷情緒。<br>*   <strong>資金避險與轉移：</strong> 在市場情緒低迷之際，投資人開始拋售風險較高的板塊，轉而將資金投入較為穩健的類股，例如公用事業和消費必需品。新聞中提到沃爾瑪和可口可樂股價分別上漲3.8%和0.5%。<br>*   <strong>經濟數據與預期：</strong> 儘管主要導火線是 AI 擔憂，但市場也在關注其他經濟指標。新聞提及了週三公佈的優於預期的美國就業報告，以及即將公佈的1月消費者物價指數（CPI）報告。經濟學家預計1月 CPI（整體和核心）將上漲0.3%，這可能影響聯準會的貨幣政策決策。<br><br>#### 台美貿易協定部分：<br><br>*   <strong>互惠貿易最終協定：</strong> 兩則新聞標題都提到了「台美簽署互惠貿易最終協定」。<br>*   <strong>關稅定調：</strong> 美國對台灣的關稅將被定調在15%，且不會疊加。<br>*   <strong>台灣擴大採購承諾：</strong> 台灣承諾在未來五年內擴大採購計畫。<br>    *   <strong>補充說明：</strong> 雖然此貿易協定訊息出現在標題，但在新聞內容主體中，<strong>並沒有深入探討此協定對當日股市下跌的直接影響。</strong> 報導的重點明顯集中在 AI 擔憂和市場的負面反應上。這項貿易協定的資訊，在內容上相對獨立於股市下跌的原因。<br><br>#### 兩則新聞的關聯性與差異：<br><br>*   <strong>高度相似：</strong> 兩則新聞在核心內容上高度相似，都報導了同一天的美股下跌情況，並將主因歸咎於對 AI 發展的擔憂，同時點出了主要指數的跌幅及部分科技股的表現。<br>*   <strong>內容細節差異：</strong><br>    *   <strong>第二則新聞更為詳盡：</strong> 第二則新聞在科技股的具體跌幅方面提供了更多細節，列出了多檔個股的表現。<br>    *   <strong>第二則新聞提及資金流向：</strong> 第二則新聞更明確地指出了資金從高風險板塊流向公用事業和消費必需品類股的現象。<br>    *   <strong>時間戳差異：</strong> 兩則新聞發布時間相近（05:54 vs 05:42），但第二則新聞的時間稍早，且在描述中提到「週四華爾街股市下跌」，第一則新聞則提到「今天下挫」。<br><br><strong>總體而言，這兩則新聞共同描繪了一幅因對 AI 潛在顛覆性影響的擔憂而引發的華爾街股市顯著下跌的景象，其中科技股受創最重。台美貿易協定的簽署雖然是另一項重要財經事件，但在本次股市下跌的直接原因分析中，並未被置於核心位置。</strong></div></div>
-        <div style="background:white; padding:15px; border-radius:10px; margin-bottom:15px; border-left:4px solid #4f46e5; box-shadow: 0 2px 4px rgba(0,0,0,0.05);">
-            <small style="color:#6366f1; font-weight:bold;">自由時報</small>
-            <h3 style="margin:5px 0; font-size:16px;">AI顛覆市場疑慮升高　道瓊收跌逾600點</h3>
-            <a href="https://ec.ltn.com.tw/article/breakingnews/5342093" target="_blank" style="color:#4f46e5; text-decoration:none; font-size:12px; font-weight:bold;">查看原文 →</a>
-        </div>
-        <div style="background:white; padding:15px; border-radius:10px; margin-bottom:15px; border-left:4px solid #4f46e5; box-shadow: 0 2px 4px rgba(0,0,0,0.05);">
-            <small style="color:#6366f1; font-weight:bold;">自由時報</small>
-            <h3 style="margin:5px 0; font-size:16px;">美股收盤》道瓊下殺669點！科技股遭殃 台積電ADR跌1.6％</h3>
-            <a href="https://ec.ltn.com.tw/article/breakingnews/5342092" target="_blank" style="color:#4f46e5; text-decoration:none; font-size:12px; font-weight:bold;">查看原文 →</a>
-        </div></div></body></html>
