--- start_bot.bat
@@ -1,12 +1,0 @@
-@echo off
-:: 1. 強制切換磁碟機與目錄 (假設你的專案在 D 槽就用 /d)
-cd /d "D:\node\news\"
-
-:: 2. 顯示當前路徑 (讓你確認有沒有切換成功)
-echo 目前執行路徑為: %cd%
-
-:: 3. 執行程式
-node index.js
-
-:: 4. 執行完後暫停，讓你看到結果不會直接閃退
-::pause
